src/backend/pyproject.toml
[project]
name = "fastapi_app"
version = "1.0.0"
description = "Create a RAG application with FastAPI and PostgreSQL"
dependencies = [
    "fastapi>=0.111.0,<1.0.0",
    "uvicorn>=0.30.1,<1.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "environs>=11.0.0,<15.0.0",
    "azure-identity>=1.16.1,<2.0.0",
    "aiohttp>=3.9.5,<4.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "SQLAlchemy[asyncio]>=2.0.30,<3.0.0",
    "pgvector>=0.3.0,<0.4.0",
    "openai>=1.108.1,<3.0.0",
    "anthropic>=0.40.0",
    "azure-monitor-opentelemetry>=1.6.0,<2.0.0",
    "opentelemetry-instrumentation-sqlalchemy",
    "opentelemetry-instrumentation-aiohttp-client",
    "opentelemetry-instrumentation-openai"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
